# openapi/Infinity.yaml
openapi: 3.0.3
info:
  version: 0.5.3
  title: Infinity API
  description: "A list of the currently provided Interfaces to interact with the infinity, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local voyager node main API
      port:
        default: 1633
        description: Service port provided in voyager node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local voyager node main API
      port:
        default: 1633
        description: Service port provided in voyager node config
paths:
  "/bytes":
    post:
      summary: "Upload data"
      tags:
        - Bytes
      parameters:
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityTagParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityPinParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityEncryptParameter"
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Ok
          headers:
            "infinity-tag":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityTag"
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/ReferenceResponse"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/bytes/{reference}":
    get:
      summary: "Get referenced data"
      tags:
        - Bytes
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address reference to content
      responses:
        "200":
          description: Retrieved content specified by reference
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
  "/chunks/{reference}":
    get:
      summary: "Get Chunk"
      tags:
        - Chunk
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address of chunk
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityRecoveryTargetsParameter"
      responses:
        "200":
          description: Retrieved chunk content
          headers:
            "infinity-recovery-targets":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityRecoveryTargets"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "202":
          description: chunk recovery initiated. retry after sometime.
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/chunks":
    post:
      summary: "Upload Chunk"
      tags:
        - Chunk
      parameters:
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityTagParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityPinParameter"
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Ok
          headers:
            "infinity-tag":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityTag"
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Status"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/files":
    post:
      summary: "Upload file"
      tags:
        - File
      parameters:
        - in: query
          name: name
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/FileName"
          required: false
          description: Filename
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityTagParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityPinParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityEncryptParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/ContentTypePreserved"
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Ok
          headers:
            "infinity-tag":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityTag"
            "etag":
              $ref: "InfinityCommon.yaml#/components/headers/ETag" 
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/files/{reference}":
    get:
      summary: "Get referenced file"
      tags:
        - File
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address of content
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "infinity-recovery-targets":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityRecoveryTargets"
            "ETag":
              $ref: "InfinityCommon.yaml#/components/headers/ETag"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/dirs":
    post:
      summary: "Upload a collection"
      tags:
        - Collection
      parameters:
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityTagParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityPinParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityEncryptParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityIndexDocumentParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityErrorDocumentParameter"
        - $ref: "InfinityCommon.yaml#/components/parameters/ContentTypePreserved"
      requestBody:
        content:
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Ok
          headers:
            "infinity-tag":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityTag"
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/ifi/{reference}":
    get:
      summary: "Get index document from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address of content
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "infinity-recovery-targets":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityRecoveryTargets"
          # "infinity-feed-index":
          #   $ref: "InfinityCommon.yaml#/components/headers/InfinityFeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/ifi/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - Collection
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/InfinityReference"
          required: true
          description: Infinity address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityRecoveryTargetsParameter"
      responses:
        "200":
          description: Ok
          headers:
            "infinity-recovery-targets":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityRecoveryTargets"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/tags":
    get:
      summary: Get list of tags
      tags:
        - Tag
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: The number of items to skip before starting to collect the result set.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          required: false
          description: The numbers of items to return.
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/TagsList"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
    post:
      summary: "Create Tag"
      tags:
        - Tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "InfinityCommon.yaml#/components/schemas/NewTagRequest"
      responses:
        "201":
          description: New Tag Info
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/NewTagResponse"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/tags/{uid}":
    get:
      summary: "Get Tag information using Uid"
      tags:
        - Tag
      parameters:
        - in: path
          name: uid
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/Uid"
          required: true
          description: Uid
      responses:
        "200":
          description: Tag info
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/NewTagResponse"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
    delete:
      summary: "Delete Tag information using Uid"
      tags:
        - Tag
      parameters:
        - in: path
          name: uid
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/Uid"
          required: true
          description: Uid
      responses:
        "204":
          $ref: "InfinityCommon.yaml#/components/responses/204"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
    patch:
      summary: "Update Total Count and infinity hash for a tag of an input stream of unknown size using Uid"
      tags:
        - Tag
      parameters:
        - in: path
          name: uid
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/Uid"
          required: true
          description: Uid
      requestBody:
        description: Can contain infinity hash to use for the tag
        required: false
        content:
          application/json:
            schema:
              $ref: "InfinityCommon.yaml#/components/schemas/Address"
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Status"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pin/chunks/{address}":
    parameters:
      - in: path
        name: address
        schema:
          $ref: "InfinityCommon.yaml#/components/schemas/InfinityAddress"
        required: true
        description: Infinity address of chunk
    post:
      summary: Pin chunk with given address
      tags:
        - Chunk pinning
      responses:
        "200":
          description: Pinning chunk with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
    delete:
      summary: Unpin chunk with given address
      tags:
        - Chunk pinning
      responses:
        "200":
          description: Unpinning chunk with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
    get:
      summary: Get pinning status of chunk with given address
      tags:
        - Chunk pinning
      responses:
        "200":
          description: Pinning state of chunk with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/PinningState"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
    put:
      summary: Update chunk pin counter
      tags:
        - Chunk pinning
      responses:
        "200":
          description: Pinning state of chunk with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/PinningState"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
  "/pin/chunks":
    get:
      summary: Get list of pinned chunks
      tags:
        - Chunk pinning
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
          required: false
          description: The number of items to skip before starting to collect the result set.
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          required: false
          description: The numbers of items to return.
      responses:
        "200":
          description: List of pinned chunks
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/IfiChunksPinned"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pin/bytes/{address}":
    parameters:
      - in: path
        name: address
        schema:
          $ref: "InfinityCommon.yaml#/components/schemas/InfinityAddress"
        required: true
        description: Infinity address of the bytes
    post:
      summary: Pin bytes with given address
      tags:
        - Bytes pinning
      responses:
        "200":
          description: Pinning bytes chunks with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
    delete:
      summary: Unpin bytes chunks with given address
      tags:
        - Bytes pinning
      responses:
        "200":
          description: Unpinning chunk with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
  "/pin/files/{address}":
    parameters:
      - in: path
        name: address
        schema:
          $ref: "InfinityCommon.yaml#/components/schemas/InfinityAddress"
        required: true
        description: Infinity address of the file
    post:
      summary: Pin file with given address
      tags:
        - File pinning
      responses:
        "200":
          description: Pinning file chunks with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
    delete:
      summary: Unpin file chunks with given address
      tags:
        - File pinning
      responses:
        "200":
          description: Unpinning file chunks with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
  "/pin/ifi/{address}":
    parameters:
      - in: path
        name: address
        schema:
          $ref: "InfinityCommon.yaml#/components/schemas/InfinityAddress"
        required: true
        description: Infinity address of the collection
    post:
      summary: Pin collection with given address
      tags:
        - Collection pinning
      responses:
        "200":
          description: Pinning collection chunks (and all referenced files) with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
    delete:
      summary: Unpin file chunks with given address
      tags:
        - Collection pinning
      responses:
        "200":
          description: Unpinning collection chunks (and all referenced files) with address
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/Response"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "403":
          $ref: "InfinityCommon.yaml#/components/responses/403"
        "404":
          $ref: "InfinityCommon.yaml#/components/responses/404"
        default:
          description: Default response
  "/pss/send/{topic}/{targets}":
    post:
      summary: Send to recipient or target with Postal Service for Infinity
      tags:
        - Postal Service for Infinity
      parameters:
        - in: path
          name: topic
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/PssTopic"
          required: true
          description: Topic name
        - in: path
          name: targets
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/PssTargets"
          required: true
          description: Target message address prefix. If multiple targets are specified, only one would be matched.
        - in: query
          name: recipient
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/PssRecipient"
          required: false
          description: Recipient publickey
      responses:
        "200":
          description: Subscribed to topic
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/pss/subscribe/{topic}":
    get:
      summary: Subscribe for messages on the given topic.
      tags:
        - Pss subscribe
      parameters:
        - in: path
          name: topic
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/PssTopic"
          required: true
          description: Topic name
      responses:
        "200":
          description: Returns a WebSocket with a subscription for incoming message data on the requested topic.
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
  "/soc/{owner}/{id}":
    post:
      summary: Upload single owner chunk
      tags:
        - Single owner chunk
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: id
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/HexString"
          required: true
          description: Id
        - in: query
          name: sig
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/HexString"
          required: true
          description: Signature
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityPinParameter"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "401":
          $ref: "InfinityCommon.yaml#/components/responses/401"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
...